Added more standardization
