+++ server/src/network/systems.rs
@@ -8,14 +8,11 @@ use crate::network::components::*;
 
 /*   START_TCP_SERVER FUNCTION   */
 ///Function that handles TCP connections seperately
-pub fn start_tcp_server(connections: Arc<Mutex<&TcpConnections>>) {
+pub fn start_tcp_server(mut connections: Arc<Mutex<TcpConnections>>) {
     //spawning thread to handle connections
     thread::spawn(move || {
         let tcp_listener = TcpListener::bind("127.0.0.1:80").unwrap();
         println!("Server listening on 127.0.0.1:80");
-
-        //variable to keep track of all connections
-        let mut connections = 0;
 
         //Accepting incoming connection
         for stream in tcp_listener.incoming() {
@@ -23,9 +20,10 @@ pub fn start_tcp_server(connections: Arc<Mutex<&TcpConnections>>) {
                 //checking it tcp connection was successfully established
                 Ok(stream) => {
                     println!(
-                        "Established a new TCP connection from {:?}",
+                        "Establishing a new TCP connection from {:?}",
                         stream.peer_addr()
                     );
+                    connections.add_connection(stream);
                 }
                 Err(e) => {
                     println!("Failed to accept connection: {:?}", e);
